Propagate possible values to clap for the lit --group flag
Improve autocomplete support for the `lit --group` flag by implementing
a custom TypedValueParser, such that we can let clap know about the
static strings we know are valid groups. As an added bonus we can
eliminate one level of Option<> on the flag type that we inherited from
the function-based impl of TypedValueParser we were relying on
previously.

There are no tests for the autocomplete improvements, as we don't have
any harness in place for them to begin with (maybe something we could
eventually implement).

The error message for this flag are also more uniform with other clap
options now, at the expense of losing the differentiation between the
"this string does not identify any known groups" and "this string
identifies multiple known groups" cases.