Enable autocomplete for minimal subcommand aliases
We already support shortened subcommand names via infer_subcommands, but
there is currently no support for infer_subcommands in clap's
autocomplete generation, so while we get a nice set of completion
candidates for a case like:

    $ cm configure -<TAB>

We unfortunately do not get any completion for the following (otherwise
equivalent) command-line:

    $ cm c -<TAB>

To work around this in the most common case, add an explicit
visible_alias for each subcommand with the shortest unambiguous name.
Currently this is always the first character of the subcommand, and
ideally this would always be true, but for less common subcommands we
may eventually add some prefix to disambiguate them without disturbing
the "hot path" of `cm c; cm a; cm b; cm l; cm d;`.