Make Native an implicit target-to-build and add opt-out alt flag
Generally one wants to enable at least the host target, even when
primarily developing another target. Reflect this by making the
-t/--targets-to-build option implicitly enable Native. Where before
a developer on an X86 machine working on AMDGPU would have to run:

    $ cm c -t X86 -t AMDGPU
    or
    $ cm c -t Native -t AMDGPU

Now, one can simply run:

    $ cm c -t AMDGPU

To control this behavior, an alternative (mutually exclusive) variant of
the option called -T/--targets-to-build-alt is added, which has the same
behavior as -t/--targets-to-build did before this patch.