Further improve autocompletion and tweak docs
Scrape "possible values" for more LLVM-isms using a bash/awk/sed script,
making autocomplete for -e/--enable-projects, -t/--targets-to-build, and
-T/--targets-to-build-alt useful while still allowing any value to
be passed through.

This is implemented through a common TypedValueParser implementation for
all of these along with lit -g/--group, with the rationale being that in
every case we want to give some "suggestions" without being rigid in
requiring values be one of the statically known values (as LLVM is
actively developed and forked, so we can never know the full set of
possible values for any of these options statically). For cases where
there is a common prefix like the lit "check-" group prefix we also
allow the shortest-unambiguous string to identify known values. I'm
on the fence as to whether this behavior is worth the more complicated
code and the need to describe the behavior, but it seems nice and
mirrors the infer_subcommands behavior.

Eventually scraping the "check-*" groups from all of the CMakeLists.txt
in llvm-project might be nice, but the list is rather long and obscure
entries might invalidate some of the shorthands allowed by the
infer_subcommands-like unambiguous-prefix behavior, so that is left to
figure out later.