Cleanup DEB package
Add a description to Cargo.toml, which is a pseudo-required field to
begin with, but which is also used for the cargo-deb Description field.

Correct the canonical path for the fish shell completion script.

In the README, used as the doc-string of binary and as the extended
description of the DEB, first refer to the `--help` output, rather
than a source file comment which may not be accessible to the end user.